+++ src/Entities/Subscription.php
@@ -15,6 +15,7 @@ use Paddle\SDK\Entities\Shared\BillingDetails;
 use Paddle\SDK\Entities\Shared\CollectionMode;
 use Paddle\SDK\Entities\Shared\CurrencyCode;
 use Paddle\SDK\Entities\Shared\CustomData;
+use Paddle\SDK\Entities\Shared\ImportMeta;
 use Paddle\SDK\Entities\Shared\TimePeriod;
 use Paddle\SDK\Entities\Subscription\SubscriptionDiscount;
 use Paddle\SDK\Entities\Subscription\SubscriptionItem;
@@ -51,6 +52,7 @@ class Subscription implements Entity
         public SubscriptionManagementUrls $managementUrls,
         public array $items,
         public CustomData|null $customData,
+        public ImportMeta|null $importMeta,
     ) {
     }
 
@@ -85,6 +87,7 @@ class Subscription implements Entity
                 : null,
             items: array_map(fn (array $item): SubscriptionItem => SubscriptionItem::from($item), $data['items']),
             customData: isset($data['custom_data']) ? new CustomData($data['custom_data']) : null,
+            importMeta: isset($data['import_meta']) ? ImportMeta::from($data['import_meta']) : null,
         );
     }
 }